service layer

+++ backend/payment/src/main/java/com/dbs/payment/service/CustomerService.java
@@ -1,5 +1,6 @@
 package com.dbs.payment.service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import com.dbs.payment.dto.CustomerDTO;
@@ -7,7 +8,7 @@ import com.dbs.payment.exception.CustomerNotFoundException;
 
 public interface CustomerService {
 	public List<CustomerDTO> getCustomers();
-	public CustomerDTO getCustomerById(int id) throws CustomerNotFoundException;
+	public CustomerDTO getCustomerById(BigInteger id) throws CustomerNotFoundException;
 	public String saveCustomer(CustomerDTO customerDTO);
-	public String deleteCustomerById(int id) throws CustomerNotFoundException;
+	public String deleteCustomerById(BigInteger id) throws CustomerNotFoundException;
 }
