+++ backend/payment/src/main/java/com/dbs/payment/service/EmployeeService.java
@@ -1,5 +1,6 @@
 package com.dbs.payment.service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import com.dbs.payment.dto.EmployeeDTO;
@@ -8,7 +9,7 @@ import com.dbs.payment.exception.EmployeeNotFoundException;
 
 public interface EmployeeService {
 	public List<EmployeeDTO> getEmployees();
-	public EmployeeDTO getEmployeeById(int id) throws EmployeeNotFoundException;
+	public EmployeeDTO getEmployeeById(BigInteger id) throws EmployeeNotFoundException;
 	public String saveEmployee(EmployeeDTO employeeDTO);
-	public String deleteEmployeeById(int id) throws EmployeeNotFoundException;
+	public String deleteEmployeeById(BigInteger id) throws EmployeeNotFoundException;
 }
